requirements.txt
streamlit==1.32.0
langchain==0.1.20
langchain-groq==0.1.4
langchain-community==0.0.38
langchain-huggingface==0.0.3
faiss-cpu==1.9.0
PyPDF2==3.0.1
sentence-transformers==2.7.0
python-dotenv==1.0.1
groq==0.5.0
tiktoken==0.7.0
